Add some basic links to the contribution guide
- Actually link to the source repository for this website
- Add a link to Github documentation on pull requests

I followed a link to the guide website itself from the SIGHPC-SYSPROS
workshop today, but didn't find an actual link from this site to the
soruce repo that others can use to help contribute. (Except buried in
the sidebar all the way at the bottom as part of the Jekyll infra.)

Along the way I added a link to help others learn how to open PRs, as
this is something I see as a common stumbling block for new folks. :)